# ensure resources concurrently to speed up reconciliation