Signed-off-by: Furkhat Kasymovgeniiuulu <vailodf@gmail.com>

**What this PR does / why we need it**: 

We experienced that sometimes cluster creation is slow. If I run `watch kubectl get all -n cluster-CLUSTERNAME` with newly created cluster name, then I see resources appearing somewhat like this: one secret - ~5 seconds - then another secret - ~5 seconds - then some configmap - ~5 seconds - another configmap - ~5 seconds - etc. I suspect it is related to the [Wait until the object exists in the cache](https://github.com/kubermatic/kubermatic/blob/v2.17.0/pkg/resources/reconciling/ensure.go#L87) in `EnsureNamedObject` function.
This changes improves speed roughly by 4 times in our environment.