@furkhat 2021-08-03
@kron4eg I don't have automated benchmark but for our installation we improved cluster creation speed from 9+ minutes to ~2 minutes.