@kron4eg 2021-08-03
I suggest at first to find the real source of performance degradation. I never observed such bad resources creation. In our dev cluster they are almost instance. Of course we don't have thousands of clusters in dev, but still, the process of reconciliation for single cluster doesn't involve other clusters so it shouldn't be a problem.

I can only speculate that it's kube-apiserver is slow in your setup for whatever reason (slow etcd? too much load? not really sure).