@furkhat 2021-08-04
@kron4eg as I have mentioned in PR description already, the "source of performance degragation" is in [wait until the object exists in the cache](https://github.com/kubermatic/kubermatic/blob/v2.17.0/pkg/resources/reconciling/ensure.go#L87)