@kron4eg 2021-08-04
> > I can't reproduce this issue in our cluster.
> 
> @kron4eg What would you suggest us to do? How do you proceed with bugs that you can't reproduce in your cluster?

I'd suggest why "the cache is slow" in your case. Do you have many thousands of clusters maybe? Or maybe it's kube-apiserver who fail to respond in timely manner.