@kron4eg 2021-08-04
> Would you consider merging this anyway? Or what would you suggest?

From my PoV, it complicates the codebase for no good reason. I'd suggest to find the real source of slowness. Having no other evidence that this is useful I'd restrain from merging it.